refactor: 코스 인프라 저장 로직 수정 및 패키지 구조 고도화 (#39)
