+++ src/main/kotlin/kr/wooco/woocobe/course/infrastructure/storage/repository/CourseCategoryJpaRepository.kt
@@ -1,5 +1,6 @@
-package kr.wooco.woocobe.course.infrastructure.storage
+package kr.wooco.woocobe.course.infrastructure.storage.repository
 
+import kr.wooco.woocobe.course.infrastructure.storage.entity.CourseCategoryEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface CourseCategoryJpaRepository : JpaRepository<CourseCategoryEntity, Long> {
