+++ src/main/kotlin/kr/wooco/woocobe/course/infrastructure/storage/repository/CourseCustomRepository.kt
@@ -1,7 +1,8 @@
-package kr.wooco.woocobe.course.infrastructure.storage
+package kr.wooco.woocobe.course.infrastructure.storage.repository
 
 import kr.wooco.woocobe.course.domain.model.CourseRegion
 import kr.wooco.woocobe.course.domain.model.CourseSortCondition
+import kr.wooco.woocobe.course.infrastructure.storage.entity.CourseEntity
 
 interface CourseCustomRepository {
     fun findAllByUserIdWithSort(
