+++ src/main/kotlin/kr/wooco/woocobe/course/infrastructure/storage/repository/CourseCustomRepositoryImpl.kt
@@ -1,8 +1,10 @@
-package kr.wooco.woocobe.course.infrastructure.storage
+package kr.wooco.woocobe.course.infrastructure.storage.repository
 
 import com.linecorp.kotlinjdsl.support.spring.data.jpa.repository.KotlinJdslJpqlExecutor
 import kr.wooco.woocobe.course.domain.model.CourseRegion
 import kr.wooco.woocobe.course.domain.model.CourseSortCondition
+import kr.wooco.woocobe.course.infrastructure.storage.entity.CourseCategoryEntity
+import kr.wooco.woocobe.course.infrastructure.storage.entity.CourseEntity
 import kr.wooco.woocobe.user.infrastructure.storage.UserEntity
 import org.springframework.stereotype.Repository
