+++ src/main/kotlin/kr/wooco/woocobe/course/infrastructure/storage/repository/InterestCourseJpaRepository.kt
@@ -1,5 +1,6 @@
-package kr.wooco.woocobe.course.infrastructure.storage
+package kr.wooco.woocobe.course.infrastructure.storage.repository
 
+import kr.wooco.woocobe.course.infrastructure.storage.entity.InterestCourseEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
